Refactoring
Refactored model used in bergen wrapper in a separate project so that it
can be shared between NED modules.

Features

AGDISTIS can now be invoked with plain text and a offsetmap of named
entities with the same format used by the bergen-wrapper to parse the
NER entities.

AGDISTIS now stores and outputs the DBPedia types of the disambiguated
entities.

Added logback-logstash encoder to logback.xml in
agdistis-bergen-wrapper.

Modified agdistis-bergen-wrapper to invoke AGDSTIS with plaintext and
entity occurences by default.